Improved the C++ / C# interface for passing back the scores and matching fragments. Comet will return all matching fragments (of all searched types) in non-sorted order. It is up to the end user to sort if they want too. Comet will return a ScoreWrapper that contains explicitly defined score metrics, instead of using a double[] Some minor conversions from long to comet_fileoffset_t to prevent losing data during copying.
git-svn-id: https://svn.code.sf.net/p/comet-ms/code@1261 299d2d02-8f3b-485e-a07e-71e8e5e16577

+++ trunk/comet-ms/CometSearch/CometWriteTxt.cpp
@@ -468,7 +468,7 @@ void CometWriteTxt::PrintResults(int iWhichQuery,
          bool bPrintDecoyPrefix = false;
          std::vector<ProteinEntryStruct>::iterator it;
 
-         int iNumTotProteins = 0;
+         size_t iNumTotProteins = 0;
 
          if (bDecoy)
          {  
@@ -522,7 +522,7 @@ void CometWriteTxt::PrintResults(int iWhichQuery,
             }
          }
 
-         fprintf(fpout, "\t%d\t", iNumTotProteins);
+         fprintf(fpout, "\t%zu\t", iNumTotProteins);
 
          // encoded modifications
          PrintModifications(fpout, pOutput, iWhichResult);
